bot_admin/requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
bleach==6.2.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
Deprecated==1.2.15
dj-database-url==2.3.0
Django==5.1.4
django-gunicorn==0.1.1
django-storages==1.14.3
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-js-asset==2.2.0
django-summernote==0.8.20.0
django-tailwind==3.8.0
django-unfold==0.44.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
idna==3.10
Jinja2==3.1.5
mysqlclient==2.2.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pikepdf==9.5.1
pillow==11.1.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
gunicorn==22.0.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.8.2
wrapt==1.17.1
social-auth-app-django==5.4.0
